Tipo de dato float

+++ JavaApplication43/src/javaapplication43/JavaApplication43.java
@@ -22,7 +22,7 @@ public class JavaApplication43 {
         System.out.println("ingrese su nombre: ");
         String nombre = leer.next();
         System.out.println("Hola " +nombre+  ", ingrese un numero : ");
-        int a= leer.nextInt();
+        float a= leer.nextInt();
         //condicional if
         
         if(a>10) {
